extract class to use the feedbacks in the EventMainFragment
